move content of service in healthCertificateService due to cycle dependency
